apps/lactation_curves/pyproject.toml
[project]
name = "lactation-curves-app"
version = "0.1.0"
description = "FastAPI backend for lactation curve analysis"
requires-python = ">= 3.12, < 3.13"
dependencies = [
    "azure-functions>=1.21.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "lactationcurve>=0.1.0",
    "pandas>=2.0.0",
]

[dependency-groups]
dev = ["ruff>=0.1.0", "basedpyright>=1.37.1"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
